+++ R/tm_outliers.R
@@ -21,8 +21,6 @@
 #' - `box_plot` (`ggplot`)
 #' - `density_plot` (`ggplot`)
 #' - `cumulative_plot` (`ggplot`)
-#' - `table` (`ElementaryTable` created with [rtables::df_to_tt()])
-#'   - The decorated table is only shown in the reporter as it is presented as an interactive `DataTable` in the module.
 #'
 #' A Decorator is applied to the specific output using a named list of `teal_transform_module` objects.
 #' The name of this list corresponds to the name of the output to which the decorator is applied.
@@ -34,8 +32,7 @@
 #'    decorators = list(
 #'      box_plot = teal_transform_module(...), # applied only to `box_plot` output
 #'      density_plot = teal_transform_module(...), # applied only to `density_plot` output
-#'      cumulative_plot = teal_transform_module(...), # applied only to `cumulative_plot` output
-#'      table = teal_transform_module(...) # applied only to `table` output
+#'      cumulative_plot = teal_transform_module(...) # applied only to `cumulative_plot` output
 #'    )
 #' )
 #' ```
@@ -198,8 +195,7 @@ tm_outliers <- function(label = "Outliers Module",
   checkmate::assert_multi_class(pre_output, c("shiny.tag", "shiny.tag.list", "html"), null.ok = TRUE)
   checkmate::assert_multi_class(post_output, c("shiny.tag", "shiny.tag.list", "html"), null.ok = TRUE)
 
-  available_decorators <- c("box_plot", "density_plot", "cumulative_plot", "table")
-  assert_decorators(decorators, names = available_decorators)
+  assert_decorators(decorators, names = c("box_plot", "density_plot", "cumulative_plot"))
   # End of assertions
 
   # Make UI args
@@ -369,7 +365,6 @@ ui_outliers <- function(id, ...) {
           decorators = select_decorators(args$decorators, "cumulative_plot")
         )
       ),
-      ui_decorate_teal_data(ns("d_table"), decorators = select_decorators(args$decorators, "table")),
       bslib::accordion_panel(
         title = "Plot settings",
         selectInput(
@@ -754,7 +749,10 @@ srv_outliers <- function(id, data, reporter, filter_panel_api, outlier_var,
       }
 
       # Generate decoratable object from data
-      qenv <- within(qenv, table <- rtables::df_to_tt(summary_data))
+      qenv <- within(qenv, {
+        table <- rtables::df_to_tt(summary_data)
+        table
+      })
 
       if (length(categorical_var) > 0 && nrow(qenv[["ANL_OUTLIER"]]) > 0) {
         shinyjs::show("order_by_outlier")
@@ -1058,14 +1056,7 @@ srv_outliers <- function(id, data, reporter, filter_panel_api, outlier_var,
       c(box_plot_q, density_plot_q, cumulative_plot_q)
     )
 
-    decorated_final_q_no_table <- reactive(decorated_q[[req(current_tab_r())]]())
-
-    decorated_final_q <- srv_decorate_teal_data(
-      "d_table",
-      data = decorated_final_q_no_table,
-      decorators = select_decorators(decorators, "table"),
-      expr = table
-    )
+    decorated_final_q <- reactive(decorated_q[[req(current_tab_r())]]())
 
     summary_table_r <- reactive({
       q <- req(decorated_final_q())